myrio-py/requirements.txt
anyio==4.9.0
biopython==1.85
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
docutils==0.21.2
fonttools==4.58.5
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
jedi==0.19.2
kiwisolver==1.4.8
loro==1.5.2
marimo==0.14.10
markdown==3.8.2
matplotlib==3.10.3
narwhals==1.46.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
polars==1.31.0
psutil==7.0.0
pygments==2.19.2
pymdown-extensions==10.16
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
tomlkit==0.13.3
tzdata==2025.2
uvicorn==0.35.0
websockets==15.0.1
